Tested communication for create and join room.
It was known that some cards have json encoding issue. It happends to be
a utf8 quote (’) character instead of normal quote ('). I attemp to
change the database character-set to utf8, but I didn't want to change
the connection in database.php. So I manullly changed all the quote. It
was like 5 of them.